Move server and api into server pkg
This allows users to easily embed the
skydns server in their applications or allow
easier mocking in their tests.
