+++ server/externalapi.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"bytes"
